fix(fe): move SubmissionDetailAdmin
