+++ apps/frontend/app/admin/contest/[contestId]/_components/SubmissionDetailAdmin.tsx
