+++ apps/frontend/app/admin/contest/[contestId]/participant/[userId]/_components/SubmissionTable.tsx
@@ -5,7 +5,7 @@ import DataTableFallback from '@/app/admin/_components/table/DataTableFallback'
 import DataTablePagination from '@/app/admin/_components/table/DataTablePagination'
 import DataTableProblemFilter from '@/app/admin/_components/table/DataTableProblemFilter'
 import DataTableRoot from '@/app/admin/_components/table/DataTableRoot'
-import SubmissionDetailAdmin from '@/app/admin/contest/[contestId]/(overall)/@tabs/submission/_components/SubmissionDetailAdmin'
+import SubmissionDetailAdmin from '@/app/admin/contest/[contestId]/_components/SubmissionDetailAdmin'
 import { Dialog, DialogContent } from '@/components/shadcn/dialog'
 import { GET_CONTEST_SUBMISSION_SUMMARIES_OF_USER } from '@/graphql/contest/queries'
 import { useSuspenseQuery } from '@apollo/client'
